split models out of unity-bridge
